requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
gunicorn==20.1.0
psycopg2==2.9.1
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4